requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
structlog>=24.1.0
prometheus-client>=0.20.0
python-dotenv>=1.0.0
tenacity>=8.2.3

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
